更新springboot版本

+++ src/main/java/cn/cebest/framework/conf/WebMvcConfig.java
@@ -5,10 +5,12 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+
 import javax.script.ScriptEngine;
 import javax.script.ScriptEngineManager;
 import javax.script.ScriptException;
 import javax.servlet.MultipartConfigElement;
+
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.web.servlet.MultipartConfigFactory;
@@ -23,54 +25,53 @@ import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+
 import lombok.extern.slf4j.Slf4j;
 
-
 /**
- *  webmvc配置
-  * @author maming  
-  * @date 2018年8月30日
+ * webmvc配置
+ * 
+ * @author maming
+ * @date 2018年8月30日
  */
 
 @Slf4j
 @Configuration
-@ConditionalOnProperty(prefix="gc.webmvc", value="enable", matchIfMissing = true)
-public class WebMvcConfig extends WebMvcConfigurerAdapter{
-	
+@ConditionalOnProperty(prefix = "gc.webmvc", value = "enable", matchIfMissing = true)
+public class WebMvcConfig extends WebMvcConfigurerAdapter {
+
 	/**
 	 * 存放访问的url及要跳转的视图
 	 */
 	@Value("#{${gc.webmvc.url2views:null}}")
-	private Map<String,String> url2views;
-	
+	private Map<String, String> url2views;
+
 	/**
 	 * 文件上传最大限制
 	 */
 	@Value("${gc.webmvc.file.maxUploadSize:10 * 1024 * 1024}")
 	private String maxUploadSize;
-	
+
 	/**
 	 * 静态资源
 	 */
 	@Value("#{${gc.webmvc.static.url2locs:null}}")
-	private Map<String,String> url2locs;
-	
+	private Map<String, String> url2locs;
+
 	/**
 	 * 拦截器
 	 */
 	@Value("#{${gc.webmvc.interceptor.pattern2class:null}}")
-	private Map<String,String> pattern2class;
+	private Map<String, String> pattern2class;
 
-	
 	/**
 	 * 使用@Value需要的bean
 	 */
 	@Bean
-	public static PropertySourcesPlaceholderConfigurer propertyConfigure(){
+	public static PropertySourcesPlaceholderConfigurer propertyConfigure() {
 		return new PropertySourcesPlaceholderConfigurer();
 	}
-	
-	
+
 	/**
 	 * 无逻辑视图跳转
 	 */
@@ -84,29 +85,28 @@ public class WebMvcConfig extends WebMvcConfigurerAdapter{
 			}
 		}
 	}
-	
-	
+
 	/**
 	 * 文件上传相关配置
+	 * 
 	 * @return
 	 */
-	
-	 @Bean
-	    public MultipartConfigElement configElement() {
-	        MultipartConfigFactory factory = new MultipartConfigFactory();
-	        ScriptEngineManager manager = new ScriptEngineManager();
-			 ScriptEngine se = manager.getEngineByName("js");
-			 try {
-		        Double  maxSize = (Double) se.eval(maxUploadSize);
-		        factory.setMaxFileSize(maxSize.longValue());
-		        factory.setMaxRequestSize(maxSize.longValue());
-			 } catch (ScriptException e) {
-				log.error("执行表达式异常", e);
-			}
-	        return factory.createMultipartConfig();
-	    }
-	
-	
+
+	@Bean
+	public MultipartConfigElement configElement() {
+		MultipartConfigFactory factory = new MultipartConfigFactory();
+		ScriptEngineManager manager = new ScriptEngineManager();
+		ScriptEngine se = manager.getEngineByName("js");
+		try {
+			Integer maxSize = (Integer) se.eval(maxUploadSize);
+			factory.setMaxFileSize(maxSize.longValue());
+			factory.setMaxRequestSize(maxSize.longValue());
+		} catch (ScriptException e) {
+			log.error("执行表达式异常", e);
+		}
+		return factory.createMultipartConfig();
+	}
+
 	/**
 	 * 设置访问静态资源
 	 */
@@ -121,8 +121,7 @@ public class WebMvcConfig extends WebMvcConfigurerAdapter{
 			}
 		}
 	}
-	
-	
+
 	/**
 	 * 配置拦截器
 	 */
@@ -134,17 +133,16 @@ public class WebMvcConfig extends WebMvcConfigurerAdapter{
 				String className = pattern2class.get(pattern);
 				try {
 					Class<?> c = Class.forName(className);
-					if(org.springframework.web.servlet.HandlerInterceptor.class.isAssignableFrom(c)){
+					if (org.springframework.web.servlet.HandlerInterceptor.class.isAssignableFrom(c)) {
 						registry.addInterceptor((HandlerInterceptor) c.newInstance()).addPathPatterns(pattern);
 					}
 				} catch (Exception e) {
 					log.error("添加拦截器异常", e);
-				} 
+				}
 			}
 		}
 	}
-	
-	
+
 	/**
 	 * 重写HttpMessageConverter
 	 */
@@ -152,13 +150,12 @@ public class WebMvcConfig extends WebMvcConfigurerAdapter{
 	public void configureMessageConverters(List<HttpMessageConverter<?>> converters) {
 		converters.add(stringHttpMessageConverter());
 	}
-	
-	
+
 	/**
 	 * 配置消息转换器
 	 */
 	@Bean
-	StringHttpMessageConverter stringHttpMessageConverter(){
+	StringHttpMessageConverter stringHttpMessageConverter() {
 		List<MediaType> supportedMediaTypes = new ArrayList<>();
 		supportedMediaTypes.add(new MediaType("text", "plain", Charset.forName("UTF-8")));
 		supportedMediaTypes.add(new MediaType("text", "html", Charset.forName("UTF-8")));
@@ -167,6 +164,5 @@ public class WebMvcConfig extends WebMvcConfigurerAdapter{
 		stringHttpMessageConverter.setSupportedMediaTypes(supportedMediaTypes);
 		return stringHttpMessageConverter;
 	}
-	
-	
+
 }
